readmes: Fix relative links
GitHub seems to set the domainpart of all internal links, relative or
absolute, to the projects base path.
Gitea seems to only adjusts relative links in Markdown-syntax and
leaves everything else as is. This is conceptually cleaner, but broke
existing links relying on GitHubs behaviour. This should now be fixed.